module-2/studio/requirements.txt
langgraph
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
notebook
langchain-tavily
wikipedia
trustcall
langgraph-cli[inmem]
langchain-mistralai
langchain-google-genai
load_dotenv